refactor: Long으로 변경

+++ src/main/java/xyz/wavit/domain/user/dto/UserWithTotalChallengeDto.java
@@ -2,9 +2,9 @@ package xyz.wavit.domain.user.dto;
 
 import xyz.wavit.domain.user.domain.User;
 
-public record UserWithTotalChallengeDto(UserFullDto user, int totalChallengeCount) {
+public record UserWithTotalChallengeDto(UserFullDto user, Long totalChallengeCount) {
 
-    public static UserWithTotalChallengeDto from(User user, int totalChallengeCount) {
+    public static UserWithTotalChallengeDto from(User user, Long totalChallengeCount) {
         return new UserWithTotalChallengeDto(UserFullDto.from(user), totalChallengeCount);
     }
 }
